+++ crates/resvg/src/tree.rs
@@ -17,7 +17,7 @@ pub struct Group {
     pub mask: Option<Mask>,
     pub filters: Vec<crate::filter::Filter>,
     pub isolate: bool,
-    /// Group's layer bounding box in canvas coordinates.
+    /// Group's layer bounding box in object coordinates.
     pub bbox: tiny_skia::Rect,
 
     pub children: Vec<Node>,
@@ -81,10 +81,6 @@ impl Tree {
     /// Text nodes should be already converted into paths using
     /// [`usvg::TreeTextToPath::convert_text`].
     pub fn from_usvg(tree: &usvg::Tree) -> Self {
-        if tree.has_text_nodes() {
-            log::warn!("Text nodes should be already converted into paths.");
-        }
-
         let (children, layer_bbox) = convert_node(tree.root.clone());
 
         Self {
@@ -128,7 +124,7 @@ impl Tree {
 
 pub fn convert_node(node: usvg::Node) -> (Vec<Node>, Option<tiny_skia::Rect>) {
     let mut children = Vec::new();
-    let bboxes = convert_node_inner(node, &mut children);
+    let bboxes = convert_node_inner(node, None, &mut children);
     (children, bboxes.and_then(|b| b.layer.to_rect()))
 }
 
@@ -139,35 +135,45 @@ pub struct BBoxes {
     /// Just a shape/image bbox as per SVG spec.
     pub object: usvg::BBox,
 
-    /// The same as above, but transformed using object's transform.
-    pub transformed_object: usvg::BBox,
-
     /// Similar to `object`, but expanded to fit the stroke as well.
     pub layer: usvg::BBox,
 }
 
-fn convert_node_inner(node: usvg::Node, children: &mut Vec<Node>) -> Option<BBoxes> {
+fn convert_node_inner(
+    node: usvg::Node,
+    text_bbox: Option<tiny_skia::NonZeroRect>,
+    children: &mut Vec<Node>,
+) -> Option<BBoxes> {
     match &*node.borrow() {
-        usvg::NodeKind::Group(ref ugroup) => convert_group(node.clone(), ugroup, children),
-        usvg::NodeKind::Path(ref upath) => crate::path::convert(upath, children),
+        usvg::NodeKind::Group(ref ugroup) => {
+            convert_group(node.clone(), ugroup, text_bbox, children)
+        }
+        usvg::NodeKind::Path(ref upath) => crate::path::convert(upath, text_bbox, children),
         usvg::NodeKind::Image(ref uimage) => crate::image::convert(uimage, children),
-        usvg::NodeKind::Text(_) => None, // should be already converted into paths
+        usvg::NodeKind::Text(ref utext) => {
+            if let (Some(bbox), Some(flattened)) = (utext.bounding_box, &utext.flattened) {
+                convert_node_inner(flattened.clone(), Some(bbox), children)
+            } else {
+                log::warn!("Text nodes should be flattened before rendering.");
+                None
+            }
+        }
     }
 }
 
 fn convert_group(
     node: usvg::Node,
     ugroup: &usvg::Group,
+    text_bbox: Option<tiny_skia::NonZeroRect>,
     children: &mut Vec<Node>,
 ) -> Option<BBoxes> {
     let mut group_children = Vec::new();
-    let mut bboxes = match convert_children(node, &mut group_children) {
+    let mut bboxes = match convert_children(node, text_bbox, &mut group_children) {
         Some(v) => v,
         None => return convert_empty_group(ugroup, children),
     };
 
-    let (filters, filter_bbox) =
-        crate::filter::convert(&ugroup.filters, bboxes.transformed_object.to_rect());
+    let (filters, filter_bbox) = crate::filter::convert(&ugroup.filters, bboxes.object.to_rect());
 
     // TODO: figure out a nicer solution
     // Ignore groups with filters but invalid filter bboxes.
@@ -192,7 +198,6 @@ fn convert_group(
     };
 
     bboxes.object = bboxes.object.transform(ugroup.transform)?;
-    bboxes.transformed_object = bboxes.transformed_object.transform(ugroup.transform)?;
     bboxes.layer = bboxes.layer.transform(ugroup.transform)?;
 
     children.push(Node::Group(group));
@@ -222,7 +227,6 @@ fn convert_empty_group(ugroup: &usvg::Group, children: &mut Vec<Node>) -> Option
     let bboxes = BBoxes {
         // TODO: find a better solution
         object: usvg::BBox::default(),
-        transformed_object: usvg::BBox::default(),
         layer: usvg::BBox::from(layer_bbox),
     };
 
@@ -230,14 +234,16 @@ fn convert_empty_group(ugroup: &usvg::Group, children: &mut Vec<Node>) -> Option
     Some(bboxes)
 }
 
-fn convert_children(parent: usvg::Node, children: &mut Vec<Node>) -> Option<BBoxes> {
+fn convert_children(
+    parent: usvg::Node,
+    text_bbox: Option<tiny_skia::NonZeroRect>,
+    children: &mut Vec<Node>,
+) -> Option<BBoxes> {
     let mut bboxes = BBoxes::default();
 
     for node in parent.children() {
-        if let Some(bboxes2) = convert_node_inner(node, children) {
+        if let Some(bboxes2) = convert_node_inner(node, text_bbox, children) {
             bboxes.object = bboxes.object.expand(bboxes2.object);
-            bboxes.transformed_object =
-                bboxes.transformed_object.expand(bboxes2.transformed_object);
             bboxes.layer = bboxes.layer.expand(bboxes2.layer);
         }
     }