+++ crates/usvg-parser/src/style.rs
@@ -4,9 +4,9 @@
 
 use usvg_tree::{ApproxEqUlps, Color, Fill, Opacity, Paint, Stroke, StrokeMiterlimit, Units};
 
-use crate::converter::SvgColorExt;
+use crate::converter::{self, SvgColorExt};
+use crate::paint_server;
 use crate::svgtree::{AId, FromValue, SvgNode};
-use crate::{converter, paint_server};
 
 impl<'a, 'input: 'a> FromValue<'a, 'input> for usvg_tree::LineCap {
     fn parse(_: SvgNode, _: AId, value: &str) -> Option<Self> {
@@ -23,6 +23,7 @@ impl<'a, 'input: 'a> FromValue<'a, 'input> for usvg_tree::LineJoin {
     fn parse(_: SvgNode, _: AId, value: &str) -> Option<Self> {
         match value {
             "miter" => Some(usvg_tree::LineJoin::Miter),
+            "miter-clip" => Some(usvg_tree::LineJoin::MiterClip),
             "round" => Some(usvg_tree::LineJoin::Round),
             "bevel" => Some(usvg_tree::LineJoin::Bevel),
             _ => None,
@@ -186,6 +187,8 @@ fn convert_paint(
                 from_fallback(node, fallback, opacity)
             }
         }
+        // Ignore `context-fill` and `context-stroke for now
+        _ => None,
     }
 }