+++ crates/usvg-text-layout/src/lib.rs
@@ -43,36 +43,7 @@ pub trait TreeTextToPath {
 impl TreeTextToPath for usvg_tree::Tree {
     fn convert_text(&mut self, fontdb: &fontdb::Database) {
         convert_text(self.root.clone(), fontdb);
-    }
-}
-
-/// A `usvg::Text` extension trait.
-pub trait TextToPath {
-    /// Converts the text node into path(s).
-    ///
-    /// `absolute_ts` is node's absolute transform. Used primarily during text-on-path resolving.
-    fn convert(&self, fontdb: &fontdb::Database, absolute_ts: Transform) -> Option<Node>;
-}
-
-impl TextToPath for Text {
-    fn convert(&self, fontdb: &fontdb::Database, absolute_ts: Transform) -> Option<Node> {
-        let (new_paths, bbox) = text_to_paths(self, fontdb, absolute_ts)?;
-
-        // Create a group will all paths that was created during text-to-path conversion.
-        let group = Node::new(NodeKind::Group(Group {
-            id: self.id.clone(),
-            transform: self.transform,
-            ..Group::default()
-        }));
-
-        let rendering_mode = resolve_rendering_mode(self);
-        for mut path in new_paths {
-            fix_obj_bounding_box(&mut path, bbox);
-            path.rendering_mode = rendering_mode;
-            group.append_kind(NodeKind::Path(path));
-        }
-
-        Some(group)
+        self.calculate_abs_transforms();
     }
 }
 
@@ -84,6 +55,12 @@ fn convert_text(root: Node, fontdb: &fontdb::Database) {
             text_nodes.push(node.clone());
         }
 
+        if let NodeKind::Image(ref mut image) = *node.borrow_mut() {
+            if let ImageKind::SVG(ref mut tree) = image.kind {
+                tree.convert_text(fontdb);
+            }
+        }
+
         node.subroots(|subroot| convert_text(subroot, fontdb))
     }
 
@@ -92,19 +69,36 @@ fn convert_text(root: Node, fontdb: &fontdb::Database) {
     }
 
     for node in &text_nodes {
-        let mut new_node = None;
-        if let NodeKind::Text(ref text) = *node.borrow() {
-            let mut absolute_ts = node.parent().unwrap().abs_transform();
-            absolute_ts = absolute_ts.pre_concat(text.transform);
-            new_node = text.convert(fontdb, absolute_ts);
-        }
-
-        if let Some(new_node) = new_node {
-            node.insert_after(new_node);
+        let absolute_ts = node.parent().unwrap().abs_transform();
+        if let NodeKind::Text(ref mut text) = *node.borrow_mut() {
+            if let Some((node, bbox)) = convert_node(text, fontdb, absolute_ts) {
+                text.bounding_box = Some(bbox);
+                text.flattened = Some(node);
+            }
         }
     }
+}
 
-    text_nodes.iter().for_each(|n| n.detach());
+fn convert_node(
+    text: &Text,
+    fontdb: &fontdb::Database,
+    absolute_ts: Transform,
+) -> Option<(Node, NonZeroRect)> {
+    let (new_paths, bbox) = text_to_paths(text, fontdb, absolute_ts)?;
+
+    let group = Node::new(NodeKind::Group(Group {
+        id: text.id.clone(),
+        ..Group::default()
+    }));
+
+    let rendering_mode = resolve_rendering_mode(text);
+    for mut path in new_paths {
+        fix_obj_bounding_box(&mut path, bbox);
+        path.rendering_mode = rendering_mode;
+        group.append_kind(NodeKind::Path(path));
+    }
+
+    Some((group, bbox))
 }
 
 trait DatabaseExt {
@@ -458,7 +452,7 @@ fn text_to_paths(
     text_node: &Text,
     fontdb: &fontdb::Database,
     abs_ts: Transform,
-) -> Option<(Vec<Path>, Rect)> {
+) -> Option<(Vec<Path>, NonZeroRect)> {
     let mut fonts_cache: FontsCache = HashMap::new();
     for chunk in &text_node.chunks {
         for span in &chunk.spans {
@@ -492,10 +486,9 @@ fn text_to_paths(
         apply_word_spacing(chunk, &mut clusters);
         apply_length_adjust(chunk, &mut clusters);
         let mut curr_pos = resolve_clusters_positions(
+            text_node,
             chunk,
             char_offset,
-            &text_node.positions,
-            &text_node.rotate,
             text_node.writing_mode,
             abs_ts,
             &fonts_cache,
@@ -560,12 +553,8 @@ fn text_to_paths(
                 }
             }
 
-            if let Some(path) = convert_span(span, &mut clusters, span_ts) {
-                // Use `text_bbox` here and not `path.data.bbox()`.
-                if let Some(r) = path.text_bbox {
-                    bbox = bbox.expand(r);
-                }
-
+            if let Some((path, span_bbox)) = convert_span(span, &mut clusters, span_ts) {
+                bbox = bbox.expand(span_bbox);
                 new_paths.push(path);
             }
 
@@ -596,7 +585,7 @@ fn text_to_paths(
         last_y = y + curr_pos.1;
     }
 
-    let bbox = bbox.to_rect()?;
+    let bbox = bbox.to_non_zero_rect()?;
     Some((new_paths, bbox))
 }
 
@@ -653,7 +642,7 @@ fn convert_span(
     span: &TextSpan,
     clusters: &mut [OutlinedCluster],
     text_ts: Transform,
-) -> Option<Path> {
+) -> Option<(Path, NonZeroRect)> {
     let mut path_builder = tiny_skia_path::PathBuilder::new();
     let mut bboxes_builder = tiny_skia_path::PathBuilder::new();
 
@@ -704,19 +693,19 @@ fn convert_span(
         fill.rule = FillRule::NonZero;
     }
 
+    let bbox = bboxes.compute_tight_bounds()?.to_non_zero_rect()?;
+
     let path = Path {
         id: String::new(),
-        transform: Transform::default(),
         visibility: span.visibility,
         fill,
         stroke: span.stroke.clone(),
         paint_order: span.paint_order,
         rendering_mode: ShapeRendering::default(),
-        text_bbox: Some(bboxes.bounds().to_non_zero_rect()?),
         data: Arc::new(path),
     };
 
-    Some(path)
+    Some((path, bbox))
 }
 
 fn collect_decoration_spans(span: &TextSpan, clusters: &[OutlinedCluster]) -> Vec<DecorationSpan> {
@@ -798,7 +787,7 @@ fn convert_decoration(
 /// By the SVG spec, `tspan` doesn't have a bbox and uses the parent `text` bbox.
 /// Since we converted `text` and `tspan` to `path`, we have to update
 /// all linked paint servers (gradients and patterns) too.
-fn fix_obj_bounding_box(path: &mut Path, bbox: Rect) {
+fn fix_obj_bounding_box(path: &mut Path, bbox: NonZeroRect) {
     if let Some(ref mut fill) = path.fill {
         if let Some(new_paint) = paint_server_to_user_space_on_use(fill.paint.clone(), bbox) {
             fill.paint = new_paint;
@@ -817,28 +806,27 @@ fn fix_obj_bounding_box(path: &mut Path, bbox: Rect) {
 /// Creates a deep copy of a selected paint server and returns its ID.
 ///
 /// Returns `None` if a paint server already uses `UserSpaceOnUse`.
-fn paint_server_to_user_space_on_use(paint: Paint, bbox: Rect) -> Option<Paint> {
+fn paint_server_to_user_space_on_use(paint: Paint, bbox: NonZeroRect) -> Option<Paint> {
     if paint.units() != Some(Units::ObjectBoundingBox) {
         return None;
     }
 
     // TODO: is `pattern` copying safe? Maybe we should reset id's on all `pattern` children.
     // We have to clone a paint server, in case some other element is already using it.
-    // If not, the `convert` module will remove unused defs anyway.
 
     // Update id, transform and units.
-    let ts = Transform::from_bbox(bbox.to_non_zero_rect()?);
+    let ts = Transform::from_bbox(bbox);
     let paint = match paint {
         Paint::Color(_) => paint,
         Paint::LinearGradient(ref lg) => {
             let transform = lg.transform.post_concat(ts);
             Paint::LinearGradient(Arc::new(LinearGradient {
-                id: String::new(),
                 x1: lg.x1,
                 y1: lg.y1,
                 x2: lg.x2,
                 y2: lg.y2,
                 base: BaseGradient {
+                    id: String::new(),
                     units: Units::UserSpaceOnUse,
                     transform,
                     spread_method: lg.spread_method,
@@ -849,13 +837,13 @@ fn paint_server_to_user_space_on_use(paint: Paint, bbox: Rect) -> Option<Paint> 
         Paint::RadialGradient(ref rg) => {
             let transform = rg.transform.post_concat(ts);
             Paint::RadialGradient(Arc::new(RadialGradient {
-                id: String::new(),
                 cx: rg.cx,
                 cy: rg.cy,
                 r: rg.r,
                 fx: rg.fx,
                 fy: rg.fy,
                 base: BaseGradient {
+                    id: String::new(),
                     units: Units::UserSpaceOnUse,
                     transform,
                     spread_method: rg.spread_method,
@@ -1366,30 +1354,23 @@ fn find_font_for_char(
 ///
 /// Returns the last text position. The next text chunk should start from that position.
 fn resolve_clusters_positions(
+    text: &Text,
     chunk: &TextChunk,
     char_offset: usize,
-    pos_list: &[CharacterPosition],
-    rotate_list: &[f32],
     writing_mode: WritingMode,
     ts: Transform,
     fonts_cache: &FontsCache,
     clusters: &mut [OutlinedCluster],
 ) -> (f32, f32) {
     match chunk.text_flow {
-        TextFlow::Linear => resolve_clusters_positions_horizontal(
-            chunk,
-            char_offset,
-            pos_list,
-            rotate_list,
-            writing_mode,
-            clusters,
-        ),
+        TextFlow::Linear => {
+            resolve_clusters_positions_horizontal(text, chunk, char_offset, writing_mode, clusters)
+        }
         TextFlow::Path(ref path) => resolve_clusters_positions_path(
+            text,
             chunk,
             char_offset,
             path,
-            pos_list,
-            rotate_list,
             writing_mode,
             ts,
             fonts_cache,
@@ -1399,10 +1380,9 @@ fn resolve_clusters_positions(
 }
 
 fn resolve_clusters_positions_horizontal(
+    text: &Text,
     chunk: &TextChunk,
     offset: usize,
-    pos_list: &[CharacterPosition],
-    rotate_list: &[f32],
     writing_mode: WritingMode,
     clusters: &mut [OutlinedCluster],
 ) -> (f32, f32) {
@@ -1411,20 +1391,20 @@ fn resolve_clusters_positions_horizontal(
 
     for cluster in clusters {
         let cp = offset + cluster.byte_idx.code_point_at(&chunk.text);
-        if let Some(pos) = pos_list.get(cp) {
+        if let (Some(dx), Some(dy)) = (text.dx.get(cp), text.dy.get(cp)) {
             if writing_mode == WritingMode::LeftToRight {
-                x += pos.dx.unwrap_or(0.0);
-                y += pos.dy.unwrap_or(0.0);
+                x += dx;
+                y += dy;
             } else {
-                y -= pos.dx.unwrap_or(0.0);
-                x += pos.dy.unwrap_or(0.0);
+                y -= dx;
+                x += dy;
             }
-            cluster.has_relative_shift = pos.dx.is_some() || pos.dy.is_some();
+            cluster.has_relative_shift = !dx.approx_zero_ulps(4) || !dy.approx_zero_ulps(4);
         }
 
         cluster.transform = cluster.transform.pre_translate(x, y);
 
-        if let Some(angle) = rotate_list.get(cp).cloned() {
+        if let Some(angle) = text.rotate.get(cp).cloned() {
             if !angle.approx_zero_ulps(4) {
                 cluster.transform = cluster.transform.pre_rotate(angle);
                 cluster.has_relative_shift = true;
@@ -1438,11 +1418,10 @@ fn resolve_clusters_positions_horizontal(
 }
 
 fn resolve_clusters_positions_path(
+    text: &Text,
     chunk: &TextChunk,
     char_offset: usize,
     path: &TextPath,
-    pos_list: &[CharacterPosition],
-    rotate_list: &[f32],
     writing_mode: WritingMode,
     ts: Transform,
     fonts_cache: &FontsCache,
@@ -1464,10 +1443,10 @@ fn resolve_clusters_positions_path(
         chunk_offset + path.start_offset + process_anchor(chunk.anchor, clusters_length(clusters));
 
     let normals = collect_normals(
+        text,
         chunk,
         clusters,
         &path.path,
-        pos_list,
         char_offset,
         start_offset,
         ts,
@@ -1494,9 +1473,7 @@ fn resolve_clusters_positions_path(
         cluster.transform = cluster.transform.pre_rotate_at(angle, half_width, 0.0);
 
         let cp = char_offset + cluster.byte_idx.code_point_at(&chunk.text);
-        if let Some(pos) = pos_list.get(cp) {
-            dy += pos.dy.unwrap_or(0.0);
-        }
+        dy += text.dy.get(cp).cloned().unwrap_or(0.0);
 
         let baseline_shift = chunk_span_at(chunk, cluster.byte_idx)
             .map(|span| {
@@ -1517,7 +1494,7 @@ fn resolve_clusters_positions_path(
                 .pre_translate(shift.x as f32, shift.y as f32);
         }
 
-        if let Some(angle) = rotate_list.get(cp).cloned() {
+        if let Some(angle) = text.rotate.get(cp).cloned() {
             if !angle.approx_zero_ulps(4) {
                 cluster.transform = cluster.transform.pre_rotate(angle);
             }
@@ -1552,10 +1529,10 @@ struct PathNormal {
 }
 
 fn collect_normals(
+    text: &Text,
     chunk: &TextChunk,
     clusters: &[OutlinedCluster],
     path: &tiny_skia_path::Path,
-    pos_list: &[CharacterPosition],
     char_offset: usize,
     offset: f32,
     ts: Transform,
@@ -1570,9 +1547,7 @@ fn collect_normals(
 
             // Include relative position.
             let cp = char_offset + cluster.byte_idx.code_point_at(&chunk.text);
-            if let Some(pos) = pos_list.get(cp) {
-                advance += pos.dx.unwrap_or(0.0);
-            }
+            advance += text.dx.get(cp).cloned().unwrap_or(0.0);
 
             let offset = advance + half_width;